safari-600.7-branch/LayoutTests/js/dom/Promise-init-callback-receiver.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script>
description('Test Promise.');

var receiverInSloppy;
new Promise(function () {
    window.receiverInSloppy = this;
    shouldBe('receiverInSloppy', 'window');
});

var receiverInStrict;
new Promise(function () {
    'use strict';
    window.receiverInStrict = this;
    shouldBe('receiverInStrict', 'undefined');
});

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
